Make row intervals internal
